+++ mangadex_downloader/network.py
@@ -39,9 +39,21 @@ def _get_netloc(url):
     result = urllib.parse.urlparse(url)
     return result.scheme + '://' + result.netloc + result.path
 
-# Modified requests session class with __del__ handler
-# so the session will be closed properly
-class requestsMangaDexSession(requests.Session):
+# Modified requests session with ability to set timeout for each requests
+class ModifiedSession(requests.Session):
+    def __init__(self):
+        super().__init__()
+
+        self._timeout = None
+
+    def set_timeout(self, time):
+        self._timeout = time
+
+    def send(self, r, **kwargs):
+        kwargs.update({'timeout': self._timeout})
+        return super().send(r, **kwargs)
+
+class requestsMangaDexSession(ModifiedSession):
     def __init__(self, trust_env=True) -> None:
         # "Circular imports" problem
         from .config import login_cache
@@ -126,9 +138,16 @@ class requestsMangaDexSession(requests.Session):
             try:
                 resp = super().request(*args, **kwargs)
             except requests.exceptions.ConnectionError as e:
-                log.error("Failed to connect to \"%s\", reason: %s. Trying... (attempt: %s)" % (
+                log.error("Failed connect to \"%s\", reason: %s. Trying... (attempt: %s)" % (
                     _get_netloc(e.request.url),
                     str(e),
+                    attempt
+                ))
+                attempt += 1
+                continue
+            except requests.exceptions.ReadTimeout as e:
+                log.error("Failed connect to '%s', reason: Connection timed out. Trying... (attempt: %s)" % (
+                    _get_netloc(e.request.url),
                     attempt
                 ))
                 attempt += 1
@@ -455,7 +474,7 @@ class NetworkObject:
     
     def _create_requests(self):
         if self._requests is None:
-            self._requests = requests.Session()
+            self._requests = ModifiedSession()
             self._update_requests_proxy(self.proxy)
     
     @property
@@ -480,6 +499,10 @@ class NetworkObject:
 
         self.requests.mount('https://', doh)
         self.requests.mount('http://', doh)
+    
+    def set_timeout(self, time):
+        self.mangadex.set_timeout(time)
+        self.requests.set_timeout(time)
 
     def close(self):
         self._mangadex.close()