+++ docs/changelog.md
@@ -5,6 +5,7 @@
 ### New features
 
 - Added DNS-over-HTTPS support
+- Added ability to set timeout for each requests
 
 ## v1.5.0